+++ src/JSONSchema/Draft4/Failure.hs
@@ -4,7 +4,7 @@ module JSONSchema.Draft4.Failure where
 import           Import
 
 import           JSONSchema.Draft4.Schema    (Schema)
-import qualified JSONSchema.Validator.Draft4 as D4
+import qualified JSONSchema.Validator.Draft4 as VAL
 
 -- | Used to report an entire instance being invalidated, as opposed
 -- to the failure of a single validator.
@@ -15,33 +15,33 @@ data Invalid = Invalid
     } deriving (Eq, Show)
 
 data ValidatorFailure
-    = FailureMultipleOf D4.MultipleOfInvalid
-    | FailureMaximum    D4.MaximumInvalid
-    | FailureMinimum    D4.MinimumInvalid
+    = FailureMultipleOf VAL.MultipleOfInvalid
+    | FailureMaximum    VAL.MaximumInvalid
+    | FailureMinimum    VAL.MinimumInvalid
 
-    | FailureMaxLength D4.MaxLengthInvalid
-    | FailureMinLength D4.MinLengthInvalid
-    | FailurePattern   D4.PatternInvalid
+    | FailureMaxLength VAL.MaxLengthInvalid
+    | FailureMinLength VAL.MinLengthInvalid
+    | FailurePattern   VAL.PatternInvalid
 
-    | FailureMaxItems        D4.MaxItemsInvalid
-    | FailureMinItems        D4.MinItemsInvalid
-    | FailureUniqueItems     D4.UniqueItemsInvalid
-    | FailureItems           (D4.ItemsInvalid ValidatorFailure)
-    | FailureAdditionalItems (D4.AdditionalItemsInvalid ValidatorFailure)
+    | FailureMaxItems        VAL.MaxItemsInvalid
+    | FailureMinItems        VAL.MinItemsInvalid
+    | FailureUniqueItems     VAL.UniqueItemsInvalid
+    | FailureItems           (VAL.ItemsInvalid ValidatorFailure)
+    | FailureAdditionalItems (VAL.AdditionalItemsInvalid ValidatorFailure)
 
-    | FailureMaxProperties     D4.MaxPropertiesInvalid
-    | FailureMinProperties     D4.MinPropertiesInvalid
-    | FailureRequired          ()
-    | FailureDependencies      (D4.DependenciesInvalid ValidatorFailure)
-    | FailurePropertiesRelated (D4.PropertiesRelatedInvalid ValidatorFailure)
+    | FailureMaxProperties     VAL.MaxPropertiesInvalid
+    | FailureMinProperties     VAL.MinPropertiesInvalid
+    | FailureRequired          VAL.RequiredInvalid
+    | FailureDependencies      (VAL.DependenciesInvalid ValidatorFailure)
+    | FailurePropertiesRelated (VAL.PropertiesRelatedInvalid ValidatorFailure)
 
-    | FailureRef   (D4.RefInvalid ValidatorFailure)
-    | FailureEnum  D4.EnumInvalid
-    | FailureType  D4.TypeValidatorInvalid
-    | FailureAllOf (D4.AllOfInvalid ValidatorFailure)
-    | FailureAnyOf (D4.AnyOfInvalid ValidatorFailure)
-    | FailureOneOf (D4.OneOfInvalid ValidatorFailure)
-    | FailureNot   D4.NotValidatorInvalid
+    | FailureRef   (VAL.RefInvalid ValidatorFailure)
+    | FailureEnum  VAL.EnumInvalid
+    | FailureType  VAL.TypeValidatorInvalid
+    | FailureAllOf (VAL.AllOfInvalid ValidatorFailure)
+    | FailureAnyOf (VAL.AnyOfInvalid ValidatorFailure)
+    | FailureOneOf (VAL.OneOfInvalid ValidatorFailure)
+    | FailureNot   VAL.NotValidatorInvalid
     deriving (Eq, Show)
 
 -- | A description of why a schema (or one of its reference) is itself